refactor: ComponentRenderPanel, ComponentControlPanel 컴포넌트 분리
